Move tests to test_i18n.py
